refactor shared coordinator zome code to library crate
